hackathon-poc/pyproject.toml
[project]
name = "crescendai-poc"
version = "0.1.0"
description = "Proof of concept for CrescendAI hackathon"
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.9.0",
    "xai-sdk>=1.4.0",
    "uvicorn>=0.32.0",
    "starlette>=0.41.0",
    "python-dotenv>=1.0.0",
    "modal>=0.72.0",
]

[tool.uv]
package = false
